feat: configure and provide store
